+++ java-client/src/main/java/io/sphere/client/shop/model/Order.java
@@ -3,6 +3,7 @@ package io.sphere.client.shop.model;
 import org.joda.time.DateTime;
 
 import javax.annotation.Nonnull;
+import java.util.Set;
 
 /** An order is the final state of a cart, usually created after a checkout process has been completed. */
 public class Order extends LineItemContainer {
@@ -10,6 +11,7 @@ public class Order extends LineItemContainer {
     @Nonnull private OrderState orderState;
     private ShipmentState shipmentState;
     private PaymentState paymentState;
+    private Set<SyncInfo> syncInfo;
 
     // for JSON deserializer
     protected Order() {}
@@ -25,6 +27,10 @@ public class Order extends LineItemContainer {
 
     /** State of the payment for this order. Optional. */
     public PaymentState getPaymentState() { return paymentState; }
+
+    public Set<SyncInfo> getSyncInfo() {
+        return syncInfo;
+    }
 
     @Override
     public String toString() {