+++ CHANGES.md
@@ -3,6 +3,7 @@
 * add field externalId to customers
 * remove OrderUpdate().addTrackingData(trackingData) and replace it with enhanced tracking API
 * add methods to change channel roles
+* integrate syncInfo field into orders
 
 ## v0.51.0